@@ -1,5 +1,4 @@
-
-class AddDeliveryMessageToSpreeOptionValues < ActiveRecord::Migration
+class AddDeliveryMessageToSpreeOptionValues < ActiveRecord::Migration[4.2]
   def change
     add_column :spree_option_values, :delivery_message, :string
   end
