@@ -6,7 +6,7 @@ describe Spree::ProductsController, type: :controller do
   let!(:variant) { create(:master_variant, product: product) }
 
   it 'should assign @variant_id if variant_id param is present' do
-    get :show, id: product.to_param, variant_id: variant.id
+    get :show, params: { id: product.to_param, variant_id: variant.id }
     expect(assigns['variant_id'].to_i).to eq(variant.id)
   end
 
@@ -16,7 +16,7 @@ describe Spree::ProductsController, type: :controller do
       create(:base_variant, product: product) # this has option values
       create(:base_variant, product: product, option_values: [])
 
-      get :show, id: product.to_param
+      get :show, params: { id: product.to_param }
       expect(assigns['variants'].count).to eq(1)
     end
   end
